+++ src/sts/game/Villager.java
@@ -16,7 +16,7 @@ public class Villager extends GameObject
     //True if the villager has already gathered gold from the mine, false if the villager is on the way to the mine, or nearest is empty
     boolean hasGold = false;
       
-    public Villager( double x, double y, Player owner )
+    public Villager( int x, int y, Player owner )
     {
         super( x, y, 1, 1, owner );
     }
